refactor: migrate to App Router, restructure project and add global styles
