+++ cookstore-next/app/globals.css
@@ -2,7 +2,7 @@ body {
   font-family: Arial, sans-serif;
   margin: 0;
   padding: 0;
-  background-color: #f5f5f5;
+  background-color: green;
 }
 
 .header {
